+++ NEWS.md
@@ -67,3 +67,12 @@
   - Replaced trivial true kinship of unstructured simulations (`I/2`) with popkin estimates in all tests.  This was necessary so WG-biased estimates would differ from their unbiased counterparts, and helped identify the numeric stability issues of the previous `ligera_f` version.
   - Added tests that confirm that `ligera_f` is invariant to WG bias (a type of bias that always results in non-singular kinship matrices).
     - Oddly, `ligera` is not invariant to WG, and it is unclear at the moment why.
+
+# ligera 1.0.10.9000 (2022-09-18)
+
+- Added function `ligera2_f`, a version of `ligera_f` that calculates performs all calculations without an explicit kinship matrix estimate (like `ligera2` was to `ligera`).
+  - Retains four `V` versions as before, and (except for `V=1`) all and agree with each other and with `ligera_f` up to machine precision.  Runtime has not been evaluated, might not be good yet for any version (f versions remain experimental).
+- Internal function changes:
+  - `popkin_prod` added option `want_inbr` (to skip calculation of `inbr`, as it is unused in f versions; default is to calculate and return it as before).
+  - `conj_grad_scan` added options `want_inbr` (to skip calculation of `inbr`), `want_b` (to return `b`, default does not as before), and `b` (allows for providing `b` as calculated in a previous run).  Additionally `mean_kinship` is now `NA` by default and it does not have to be provided if `b` is provided.  Lastly, if `want_inbr` and `want_b` are both false, then only `Z` is returned (instead of a list with these objects and `Z`).
+- Corrected minor errors in non-f unit tests (the wrong functions or objects were being tested, though the changes were inconsequential).